fix: fix course payment - change amount type from Long to Double to handle decimal prices from CourseDetailResponse

+++ backend/src/main/java/com/ai/learning/backend/payment/dto/request/MomoRequest.java
@@ -2,7 +2,7 @@ package com.ai.learning.backend.payment.dto.request;
 
 public record MomoRequest(
         Long userId,
-        Long amount,
+        Double amount,
         String planType,
         Long courseId
 ) {}
